fixing #606
